Move back RustdocPage to `web::rustdoc`
